+++ database/factories/EloquentUserFactory.php
@@ -48,4 +48,12 @@ class EloquentUserFactory extends Factory
             'email_verified_at' => null,
         ]);
     }
+
+    public function withoutName(): static
+    {
+        return $this->state(fn(array $attributes) => [
+            'first_name' => null,
+            'last_name'  => null,
+        ]);
+    }
 }